// packages/web/src/components/charts/types.ts
export type chartOptionsType = Partial<{
    backgroundColor: string;
    borderColor: string;
    borderWidth: number;
    gridColor: string;
    legendLabelColor: string;
    pointBackgroundColor: string;
    pointBorderColor: string;
    tension: number;
    ticksColor: string;
    titleColor: string;
}>;